% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{carrion_biomass}
\alias{carrion_biomass}
\title{Carrion biomass}
\usage{
carrion_biomass(params)
}
\arguments{
\item{params}{MizerParams}
}
\value{
The carrion biomass in grams
}
\description{
Carrion biomass
}
