% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{carrion_dynamics}
\alias{carrion_dynamics}
\title{Carrion dynamics}
\usage{
carrion_dynamics(params, n, n_other, rates, dt, ...)
}
\arguments{
\item{params}{A \link{MizerParams} object}

\item{n}{A matrix of current species abundances (species x size)}

\item{n_other}{Other dynamic components. Only \code{n_other$carrion} is used.}

\item{rates}{A list of rates as returned by \code{\link[=getRates]{getRates()}}}

\item{dt}{Time step size}

\item{...}{Unused}
}
\value{
A single number giving the component biomass at next time step
}
\description{
Calculates the carrion biomass at the next timestep from the current
carrion biomass.
}
\details{
The equation for the time evolution of the carrion biomass \eqn{B} is
assumed to be of the form
\deqn{dB/dt = \tt{production} - \tt{consumption} * B}{dB/dt = production - consumption * B}
where
\itemize{
\item \code{consumption} is the mass-specific rate of consumption due to consumption, where the
encounter rate is specified by \code{params@other_params[[component]]$rho}.
\item \code{production} is the rate at which the rest of the system produces carrion
biomass. The production comes from three sources: 1) animals
killed by the fishing gear, 2) animals that have died by natural causes
other than predation, 3) discards from fishing.
}

The dynamical equation is solved analytically to
\deqn{B(t+dt) = B(t)\exp(-\tt{consumption} \cdot dt)
  +\frac{\tt{production}}{\tt{consumption}}
  (1-\exp(-\tt{consumption} \cdot dt)).}{B(t+dt)
  = B(t) exp(-consumption * dt) + production/consumption * (1 - exp(-consumption * dt)).}
This avoids the stability problems that would arise if we used the Euler
method to solve the equation numerically.
}
