% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{detritus_biomass}
\alias{detritus_biomass}
\title{Detritus biomass}
\usage{
detritus_biomass(params, n_pp = params@initial_n_pp)
}
\arguments{
\item{params}{MizerParams}
}
\value{
The detritus biomass in grams
}
\description{
The detritus is internally described by a size spectrum in order to reflect
the fact that it is available to small predators but becomes unavailable to
large predators. The total biomass is thus obtained by integrating over
the abundance density multiplied by mass.
}
