% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{detritus_biomass_consumption}
\alias{detritus_biomass_consumption}
\title{Detritus consumption rate}
\usage{
detritus_biomass_consumption(
  params,
  n_pp = params@initial_n_pp,
  rates = getRates(params)
)
}
\arguments{
\item{params}{MizerParams}

\item{rates}{A list of rates as returned by \code{\link[=getRates]{getRates()}}}

\item{n}{A matrix of current species abundances (species x size)}
}
\value{
A number giving the consumption rate in grams per year.
}
\description{
Detritus consumption rate
}
