% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{detritus_consumption}
\alias{detritus_consumption}
\title{Detritus consumption rate}
\usage{
detritus_consumption(
  params,
  n_pp = params@initial_n_pp,
  rates = getRates(params)
)
}
\arguments{
\item{params}{MizerParams}

\item{rates}{A list of rates as returned by \code{\link[=getRates]{getRates()}}}

\item{n}{A matrix of current species abundances (species x size)}
}
\value{
A number giving the consumption rate in grams per year.
}
\description{
An internal helper function. This returns the total detritus consumption rate
and is used in \code{detritus_dynamics()} to calculate the detritus abundance at
the next time step. To get the consumption rate split up by consumer species,
use \code{getDetritusConsumption()}.
}
