% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{detritus_dynamics}
\alias{detritus_dynamics}
\title{Detritus dynamics}
\usage{
detritus_dynamics(params, n, n_pp, n_other, rates, dt, ...)
}
\arguments{
\item{params}{A \link{MizerParams} object}

\item{n}{A matrix of current species abundances (species x size)}

\item{n_pp}{Vector of detritus density}

\item{n_other}{Other dynamic components. Only \code{n_other$carrion} is used.}

\item{rates}{A list of rates as returned by \code{\link[=getRates]{getRates()}}}

\item{dt}{Time step size}

\item{...}{Unused}
}
\value{
A vector giving the detritus spectrum at the next time step.
}
\description{
Calculates the detritus size spectrum at the next time step from the current
size spectrum. The size spectrum is always held at a power law with the
same exponent -- only the intercept is dynamical to reflect the change in
total detritus biomass.
}
\details{
The time evolution of the detritus biomass \eqn{B} is described by
\deqn{dB/dt = \tt{production} - \tt{consumption} * B + \tt{external}}{dB/dt = production - consumption * B + external}
where
\itemize{
\item \code{consumption} is the mass-specific rate of consumption.
\item \code{production} is the rate at which the rest of the system produces
detritus biomass.
}

The dynamical equation is solved analytically to
\deqn{B(t+dt) = B(t)\exp(-\tt{consumption} \cdot dt)
  +\frac{\tt{production}}{\tt{consumption}}
  (1-\exp(-\tt{consumption} \cdot dt)).}{B(t+dt)
  = B(t) exp(-consumption * dt)
  +production/consumption * (1 - exp(-consumption * dt)).}
This avoids the stability problems that would arise if we used the Euler
method to solve the equation numerically.
}
