% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{detritus_lifetime}
\alias{detritus_lifetime}
\alias{detritus_lifetime<-}
\title{Expected detritus lifetime}
\usage{
detritus_lifetime(params)

detritus_lifetime(params) <- value
}
\arguments{
\item{params}{A MizerParams object}

\item{value}{A number with the new value for the lifetime}
}
\value{
The number giving the expected lifetime in years.
}
\description{
The expected detritus lifetime is defined as the inverse of the
mass-specific detritus consumption rate.
}
