% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{gearMort}
\alias{gearMort}
\title{Excess gear mortality rate}
\usage{
gearMort(params, f_mort)
}
\arguments{
\item{params}{A MizerParams object}

\item{f_mort}{The array of fishing mortality rates as calculated by
\code{getFMort()}.}
}
\value{
A named two dimensional array (species x size) with the excess gear
mortality rates.
}
\description{
The excess gear mortality rate is the difference between the total
gear mortality (as given by the species
parameter \code{gear_mort}) and the fishing mortality) if this is positive.
If the total gear mortality does not exceed the fishing mortality then the
excess gear mortality is zero.
}
