% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizerShelf-class.R
\name{getBiomass.mizerShelfSim}
\alias{getBiomass.mizerShelfSim}
\title{Get biomass of species and components through time for a shelf model}
\usage{
\method{getBiomass}{mizerShelfSim}(object, ...)
}
\arguments{
\item{object}{A \code{mizerShelfSim} object.}

\item{...}{Passed to the base \code{\link[mizer:getBiomass]{mizer::getBiomass()}} method.}
}
\value{
An \code{ArraySpeciesByTime} matrix (time x species/component) with
species biomasses followed by Detritus and other component biomasses.
}
\description{
Extends \code{\link[mizer:getBiomass]{mizer::getBiomass()}} for \code{mizerShelfSim} objects by adding the
detritus (resource spectrum) and any scalar other-component biomasses (e.g.
carrion) as extra columns.
}
