% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{getCarrionConsumption}
\alias{getCarrionConsumption}
\title{Get carrion consumption rates}
\usage{
getCarrionConsumption(params, ...)
}
\arguments{
\item{params}{A \link[mizer:MizerParams]{MizerParams} object.}

\item{...}{Unused}
}
\value{
A named vector with the consumption rates from all species and
decomposition.
}
\description{
Returns a named vector with one entry for each species giving the rate in
grams/year at which that species consumes carrion, plus a \code{decompose} entry
for bacterial decomposition.
}
\seealso{
\code{\link[=getCarrionProduction]{getCarrionProduction()}}, \code{\link[=carrion_dynamics]{carrion_dynamics()}}, \code{\link[=getDetritusConsumption]{getDetritusConsumption()}}
}
