% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{getCarrionProduction}
\alias{getCarrionProduction}
\title{Carrion production rate}
\usage{
getCarrionProduction(params, n = params@initial_n, rates = getRates(params))
}
\arguments{
\item{params}{MizerParams}

\item{n}{A matrix of current species abundances (species x size)}

\item{rates}{A list of rates as returned by \code{\link[=getRates]{getRates()}}}
}
\value{
A vector with named entries "external",
"gear_mort" and "discards", each given the rate at which carrion biomass
is produced by these sources in grams per year.
}
\description{
This is the rate in grams/year at which the rest of the system produces carrion
biomass. The production comes from three sources:
}
\details{
\enumerate{
\item animals that have died by natural causes other than predation ("ext_mort"),
\item animals killed by the fishing gear ("gear_mort"),
\item discards from fishing ("discards").
}

The function returns a vector with the individual contributions. These
can be summed with \code{sum()} to get the total production rate.
}
\seealso{
getCarrionConsumption carrion_dynamics getDetritusProduction
}
