% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newDetritusCarrionParams.R
\name{newDetritusCarrionParams}
\alias{newDetritusCarrionParams}
\title{Create new mizer model with detritus and carrion components}
\usage{
newDetritusCarrionParams(w_min_detritus = NA, w_max_detritus = 1, n = 0.7, ...)
}
\arguments{
\item{w_min_detritus}{Minimum size of detritus in grams}

\item{w_max_detritus}{Maximum size of detritus in grams}

\item{n}{Growth exponent (also used as metabolic exponent p)}

\item{...}{Extra parameters to be passed to \code{\link[=newMultispeciesParams]{newMultispeciesParams()}}}

\item{species_params}{A species parameter data frame}
}
\value{
A MizerParams object
}
\description{
Create new mizer model with detritus and carrion components
}
