% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NWMed_catch}
\alias{NWMed_catch}
\title{Catch size distribution for the NW Mediterranean shelf around Blanes}
\format{
A data frame with 208 rows and 4 columns:
\describe{
\item{species}{Species name}
\item{length}{Start of length bin in cm}
\item{dl}{Width of length bin in cm. The bin goes from \code{length} to
\code{length + dl}.}
\item{catch}{Number of fish observed in the catch in this size bin for this
species}
}
}
\source{
Silvia de Juan, IMEDEA (UIB-CSIC)
}
\usage{
NWMed_catch
}
\description{
This contains size observations from landings of 9 species:
Shortfin squid, Angler fish, Poor cod, Hake, Blue whiting,
Striped red mullet, Horse mackerel, Red mullet and Starfish.
}
\keyword{datasets}
