% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{plotCarrionConsumption}
\alias{plotCarrionConsumption}
\title{Plot carrion consumption rates}
\usage{
plotCarrionConsumption(params)
}
\arguments{
\item{params}{MizerParams}
}
\value{
A pie chart.
}
\description{
Plot carrion consumption rates
}
