% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{plotCarrionProduction}
\alias{plotCarrionProduction}
\title{Plot carrion production rates}
\usage{
plotCarrionProduction(params)
}
\arguments{
\item{params}{MizerParams}
}
\value{
A pie chart.
}
\description{
Plot carrion production rates
}
