% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{plotDetritusProduction}
\alias{plotDetritusProduction}
\title{Plot detritus production rates}
\usage{
plotDetritusProduction(params)
}
\arguments{
\item{params}{MizerParams}
}
\value{
A pie chart.
}
\description{
Plot detritus production rates
}
