% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotYieldMinusDiscards}
\alias{plotYieldMinusDiscards}
\title{Plot yield minus discards}
\usage{
plotYieldMinusDiscards(
  sim,
  sim2,
  species = NULL,
  total = FALSE,
  log = TRUE,
  highlight = NULL,
  return_data = FALSE,
  ...
)
}
\arguments{
\item{sim}{An \code{\link{Sim}} object.}

\item{sim2}{Optional second \code{\link{Sim}} object.}

\item{species}{Species to plot.}

\item{total}{Logical; should total yield be included in plot?}

\item{log}{Logical; should the y-axis be log-transformed?}

\item{highlight}{Name of species to highlight in plot.}

\item{return_data}{Logical; should the underlying data be returned?}

\item{...}{Other arguments passed to \code{\link[Sims]{getYield}}.}
}
\value{
If \code{return_data = TRUE}, a data frame of the underlying
data is returned.
}
\description{
Plot yield minus discards
}
\examples{
plotYieldMinusDiscards(sim_instance)
}
