% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{removeSpecies}
\alias{removeSpecies}
\alias{removeSpecies.mizerShelf}
\title{Remove some species from a shelf model}
\usage{
\method{removeSpecies}{mizerShelf}(params, species, ...)
}
\arguments{
\item{params}{A \code{mizerShelf} params object.}

\item{species}{The species to be removed. A vector of species names, or a
numeric vector of species indices, or a logical vector indicating for each
species whether it is to be removed (TRUE) or not.}

\item{...}{Passed to \code{\link[mizer:removeSpecies]{mizer::removeSpecies()}}.}
}
\value{
A \code{mizerShelf} params object with fewer species.
}
\description{
Extends \code{\link[mizer:removeSpecies]{mizer::removeSpecies()}} for \code{mizerShelf} objects: after removing
the species it also removes the corresponding rows from the carrion
encounter rate matrix \code{rho}.
}
\examples{
params <- NWMed_params
species_params(params)$species
params <- removeSpecies(params, c("Poor cod", "Horse mackerel"))
species_params(params)$species
}
