% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{rescale_carrion}
\alias{rescale_carrion}
\title{Rescale carrion biomass without changing anything else}
\usage{
rescale_carrion(params, factor)
}
\arguments{
\item{params}{A MizerParams object}

\item{factor}{A number}
}
\value{
An updated MizerParams object
}
\description{
This multiplies the carrion biomass by a factor and divides the
interaction between all species and the detritus by the same
factor, so as to keep the total consumption of detritus unchanged.
It also divides the mass-specific rate of decomposition by the same
factor so that the total decomposition rate stays the same.
}
