% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{rescaleComponents}
\alias{rescaleComponents}
\title{Rescale carrion and detritus biomass without changing anything else}
\usage{
rescaleComponents(params, carrion_factor = 1, detritus_factor = 1)
}
\arguments{
\item{params}{A MizerParams object}

\item{carrion_factor}{A number by which to multiply the carrion biomass}

\item{detritus_factor}{A number by which to multiply the detritus biomass}
}
\value{
An updated MizerParams object
}
\description{
A convenience wrapper that calls \code{\link[=rescale_carrion]{rescale_carrion()}} and
\code{\link[=rescale_detritus]{rescale_detritus()}}.
}
