% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhoControl.R
\name{rhoControl}
\alias{rhoControl}
\alias{rhoControlUI}
\title{Controlling the carrion encounter rate in the tuning gadget}
\usage{
rhoControl(input, output, session, params, params_old, flags, ...)

rhoControlUI(p, input)
}
\arguments{
\item{input}{Reactive holding the inputs}

\item{output}{Reactive holding the outputs}

\item{session}{Shiny session}

\item{params}{Reactive value holding updated MizerParams object}

\item{params_old}{Reactive value holding non-updated MizerParams object}

\item{flags}{Environment holding flags to skip certain observers}

\item{...}{Unused}

\item{p}{The MizerParams object currently being tuned.}
}
\description{
This control allows the user to adjust \code{rho_carrion}, the coefficient of
the species' encounter rate for carrion, for the selected species.
}
\concept{gadget controls}
