% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{seMort}
\alias{seMort}
\title{Total mortality rate in the shelf ecosystem model}
\usage{
seMort(params, n, n_pp, n_other, t, f_mort, pred_mort, ...)
}
\arguments{
\item{params}{A MizerParams object}

\item{n}{A matrix of species abundances (species x size).}

\item{n_pp}{A vector of the resource abundance by size}

\item{n_other}{A list of abundances for other dynamical components of the
ecosystem}

\item{t}{The time for which to do the calculation (Not used by standard
mizer rate functions but useful for extensions with time-dependent
parameters.)}

\item{f_mort}{A two dimensional array (species x size) with the fishing
mortality}

\item{pred_mort}{A two dimensional array (species x size) with the predation
mortality}

\item{...}{Unused}
}
\value{
A named two dimensional array (species x size) with the total
mortality rates.
}
\description{
Total mortality rate in the shelf ecosystem model
}
