% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhoControl.R
\name{setRho}
\alias{setRho}
\title{Set the carrion encounter rate from the \code{rho_carrion} species parameter}
\usage{
setRho(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
An updated MizerParams object
}
\description{
Set the carrion encounter rate from the \code{rho_carrion} species parameter
}
