% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{carrion_consumption_ms}
\alias{carrion_consumption_ms}
\title{Mass-specific carrion consumption rate}
\usage{
carrion_consumption_ms(params, n = params@initial_n, rates = getRates(params))
}
\arguments{
\item{params}{MizerParams}

\item{n}{A matrix of current species abundances (species x size)}

\item{rates}{A list of rates as returned by \code{\link[=getRates]{getRates()}}}
}
\value{
A number giving the mass-specific consumption rate in grams per year.
}
\description{
This includes both the consumption by fish and the decomposition by smaller
organisms.
}
\details{
This mass-specific consumption rate is used in \code{carrion_dynamics()} to
calculate the carrion biomass at the next time step. To get the
non-mass-specific consumption rate, use \code{getCarrionConsumption()}.

The consumption rate by fish is determined by
\code{params@other_params$carrion$rho}
and the decomposition rate is given by
\code{params@other_params$carrion$decompose}.
}
