% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{carrion_lifetime}
\alias{carrion_lifetime}
\alias{carrion_lifetime<-}
\title{Expected carrion lifetime}
\usage{
carrion_lifetime(params)

carrion_lifetime(params) <- value
}
\arguments{
\item{params}{A MizerParams object}

\item{value}{A number with the new value for the lifetime}
}
\value{
The number giving the expected lifetime in years.
}
\description{
The expected carrion lifetime is defined as the inverse of the
mass-specific carrion consumption rate.
}
