% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{encounter_contribution}
\alias{encounter_contribution}
\title{Contribution of unstructured components to the encounter rate}
\usage{
encounter_contribution(params, n_other, component, ...)
}
\arguments{
\item{params}{MizerParams}

\item{n_other}{Biomasses of unstructured components}

\item{component}{Name of component whose contribution is requested}
}
\value{
Array species x size
}
\description{
Contribution of unstructured components to the encounter rate
}
