% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizerShelf-class.R
\name{getBiomass}
\alias{getBiomass}
\alias{getBiomass.mizerShelf}
\title{Get biomass of species and components for a shelf model}
\usage{
\method{getBiomass}{mizerShelf}(object, ...)
}
\arguments{
\item{object}{A \code{mizerShelf} or \code{mizerShelfSim} object.}

\item{...}{Passed to the base \code{\link[mizer:getBiomass]{mizer::getBiomass()}} method.}
}
\value{
For \code{mizerShelf}: a named numeric vector of species/component
biomasses. For \code{mizerShelfSim}: an \code{ArrayTimeBySpecies} matrix
(time x species/component) with species biomasses followed by Detritus
and other component biomasses.
}
\description{
Extends \code{\link[mizer:getBiomass]{mizer::getBiomass()}} for \code{mizerShelf} and \code{mizerShelfSim} objects
by adding the detritus (resource spectrum) and any scalar other-component
biomasses (e.g. carrion).
}
\examples{
getBiomass(NWMed_params)
\donttest{
sim <- project(NWMed_params, t_max = 3)
getBiomass(sim)
}
}
