% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{getCarrionConsumption}
\alias{getCarrionConsumption}
\title{Get carrion consumption rates}
\usage{
getCarrionConsumption(params)
}
\arguments{
\item{params}{MizerParams}
}
\value{
A named vector with the consumption rates from all species and
decomposition.
}
\description{
This function returns a named vector with one component for each species
giving the rate in grams/year at which that species consumes carrion, as well
as a component named \code{decompose} with the rate in grams/year at which carrion
is decomposed by bacteria and other processes.
}
\seealso{
getCarrionProduction carrion_dynamics getDetritusConsumption
}
