% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{getCarrionConsumption}
\alias{getCarrionConsumption}
\title{Get carrion consumption rates}
\usage{
getCarrionConsumption(params)
}
\arguments{
\item{params}{MizerParams}
}
\value{
A named vector with the consumption rates from all species and
decomposition.
}
\description{
Get carrion consumption rates
}
