% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{getDetritusConsumption}
\alias{getDetritusConsumption}
\title{Get detritus consumption rates by consumer species}
\usage{
getDetritusConsumption(params)
}
\arguments{
\item{params}{MizerParams}
}
\value{
A named vector with the consumption rates from all species
}
\description{
This function returns a named vector with one component for each species
giving the rate in grams/year at which that species consumes detritus.
}
\seealso{
\code{\link[=getDetritusProduction]{getDetritusProduction()}}, \code{\link[=detritus_dynamics]{detritus_dynamics()}}, \code{\link[=getCarrionConsumption]{getCarrionConsumption()}}
}
