% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{getDetritusConsumption}
\alias{getDetritusConsumption}
\title{Get detritus consumption rates by consumer species}
\usage{
getDetritusConsumption(params)
}
\arguments{
\item{params}{MizerParams}
}
\value{
A named vector with the consumption rates from all species
}
\description{
Get detritus consumption rates by consumer species
}
