% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{getDetritusProduction}
\alias{getDetritusProduction}
\title{Detritus production rate}
\usage{
getDetritusProduction(
  params,
  n = params@initial_n,
  n_other = params@initial_n_other,
  rates = getRates(params)
)
}
\arguments{
\item{params}{MizerParams}

\item{n}{A matrix of current species abundances (species x size)}

\item{rates}{A list of rates as returned by \code{\link[=getRates]{getRates()}}}
}
\value{
A vector with named entries "external",
"feces" and "carrion", each given the rate at which carrion biomass
is produced by these sources in grams per year.
}
\description{
This is the rate at which the rest of the system produces detritus
biomass. The production comes from three sources:
}
\details{
\enumerate{
\item biomass not assimilated by predators ("feces"),
\item decomposing carrion ("carrion"),
\item the pelagic zone ("external").
}

The function returns a vector with the individual contributions. These
can be summed with \code{sum()} to get the total production rate.
}
