% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizerShelf-class.R
\name{mizerShelf-class}
\alias{mizerShelf-class}
\title{mizerShelf extension classes}
\description{
S3 extension classes for \link[mizer:MizerParams]{MizerParams} and \link[mizer:MizerSim]{MizerSim} that enable S3 dispatch
for shelf-specific methods such as \code{\link[=tuneSteadyState]{tuneSteadyState()}},
\code{\link[=scaleModel]{scaleModel()}},
\code{\link[=removeSpecies]{removeSpecies()}}, \code{\link[=addSpecies]{addSpecies()}}, and \code{\link[=getBiomass]{getBiomass()}}.
}
\details{
The class names are ordinary entries in the object's S3 class vector. All
extension-specific data lives in \code{other_params(params)} or in the component
parameters of the "carrion" component (see \code{\link[mizer:setComponent]{setComponent()}}).

Objects of class \code{mizerShelf} are created by \code{\link[=newDetritusCarrionParams]{newDetritusCarrionParams()}}.
Objects of class \code{mizerShelfSim} are returned automatically by \code{\link[mizer:project]{project()}}
when called on a \code{mizerShelf} params object.

No class declaration is needed. \code{\link[=newDetritusCarrionParams]{newDetritusCarrionParams()}} records the
extension on the object with \code{\link[mizer:recordExtension]{mizer::recordExtension()}} and then calls
\code{\link[mizer:coerceToExtensionClass]{mizer::coerceToExtensionClass()}}. For example, the params class vector is
\code{c("mizerShelf", "MizerParams")}; simulations created by \code{\link[mizer:project]{project()}} receive
\code{c("mizerShelfSim", "MizerSim")} automatically.
}
\keyword{internal}
