% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizerShelf-class.R
\name{mizerShelf-class}
\alias{mizerShelf-class}
\title{mizerShelf marker classes}
\description{
S4 marker subclasses of \link[mizer:MizerParams]{MizerParams} and \link[mizer:MizerSim]{MizerSim} that enable S3 dispatch
for shelf-specific methods such as \code{\link[=steady]{steady()}}, \code{\link[=scaleModel]{scaleModel()}},
\code{\link[=removeSpecies]{removeSpecies()}}, \code{\link[=addSpecies]{addSpecies()}}, and \code{\link[=getBiomass]{getBiomass()}}.
}
\details{
Objects of class \code{mizerShelf} are created by \code{\link[=newDetritusCarrionParams]{newDetritusCarrionParams()}}.
Objects of class \code{mizerShelfSim} are returned automatically by \code{\link[mizer:project]{project()}}
when called on a \code{mizerShelf} params object.
}
