% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NWMed_params}
\alias{NWMed_params}
\title{MizerParams object for the NW Mediterranean shelf model}
\format{
A MizerParams object
}
\source{
Forthcoming paper
}
\usage{
NWMed_params
}
\description{
MizerParams object for the NW Mediterranean shelf model
}
\keyword{datasets}
