% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDeath}
\alias{plotDeath}
\alias{plotlyDeath}
\title{Plot Death}
\usage{
plotDeath(object, species = NULL, proportion = TRUE, return_data = FALSE)

plotlyDeath(object, species = NULL, proportion = TRUE, ...)
}
\description{
Plot Death
}
