% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{removeSpecies}
\alias{removeSpecies}
\title{Remove some species from the model}
\usage{
removeSpecies(params, species)
}
\arguments{
\item{params}{A MizerParams object}

\item{species}{The species to be removed. A vector of species names, or a
numeric vector of species indices, or a logical vector indicating for each
species whether it is to be removed (TRUE) or not.}
}
\value{
A MizerParams object with fewer species.
}
\description{
This calls \code{mizer::removeSpecies()} and in addition removes the relevant
row from the carrion consumption array \code{rho}.
}
\examples{
params <- NWMed_params
species_params(params)$species
params <- removeSpecies(params, c("Poor cod", "Horse mackerel"))
species_params(params)$species
}
