% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steady.R
\name{steady}
\alias{steady}
\alias{steady.mizerShelf}
\title{Drive a shelf model to steady state}
\usage{
\method{steady}{mizerShelf}(
  params,
  t_max = 100,
  t_per = 1.5,
  dt = 0.1,
  tol = 0.1 * dt,
  return_sim = FALSE,
  ...
)
}
\arguments{
\item{params}{A \code{mizerShelf} params object.}

\item{t_max}{The maximum number of years to run the simulation. Default is 100.}

\item{t_per}{The simulation is broken up into shorter runs of \code{t_per} years,
after each of which we check for convergence. Default value is 1.5. This
should be chosen as an odd multiple of the timestep \code{dt} in order to be
able to detect period 2 cycles.}

\item{dt}{The time step to use in \code{project()}.}

\item{tol}{The simulation stops when the relative change in the egg
production RDI over \code{t_per} years is less than \code{tol} for every species.}

\item{return_sim}{If TRUE, the function returns the MizerSim object holding
the result of the simulation run, saved at intervals of \code{t_per}. If FALSE (default) the function returns
a MizerParams object with the "initial" slots set to the steady state.}

\item{...}{Passed to \code{\link[mizer:steady]{mizer::steady()}}, for example \code{preserve},
\code{progress_bar}, \code{info_level} or \code{method}.}
}
\value{
An updated \code{mizerShelf} object (or a \code{mizerShelfSim} when
\code{return_sim = TRUE}).
}
\description{
Extends \code{\link[mizer:steady]{mizer::steady()}} for \code{mizerShelf} objects: after the fish
abundances have converged, \code{\link[=tune_carrion_detritus]{tune_carrion_detritus()}} is called so that the
carrion and detritus components are also at steady state.
}
