% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{tune_carrion_detritus}
\alias{tune_carrion_detritus}
\title{Tune carrion and detritus to steady state}
\usage{
tune_carrion_detritus(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
An updated MizerParams object
}
\description{
This first sets the rate of decomposition of carrion so that for the given
abundances, the carrion is at steady state. It then sets the rate at which
detritus flows in from external sources (e.g. the pelagic zone) so that for
the given abundances the detritus is at steady state.
}
